[Feature][UI] front-end change from review comments
