+++ dolphinscheduler-ui/src/views/projects/task/components/node/fields/index.ts
@@ -62,5 +62,6 @@ export { useConditions } from './use-conditions'
 export { useDependent } from './use-dependent'
 export { useEmr } from './use-emr'
 export { useZeppelin } from './use-zeppelin'
+export { useNamespace } from './use-namespace'
 export { useK8s } from './use-k8s'
 export { useJupyter } from './use-jupyter'
